requirements.txt
pytest
urllib3
brotlipy
